Move check out aspects of hold shelf fulfilment to more specific test class CIRC-260
